test: simplify resource export tests
